+++ src/lib/components/forms/form/model/form-type-map.ts
@@ -7,7 +7,7 @@ import { TerraCategoryPickerComponent } from '../../../data-picker/category-pick
 import { TerraNoteEditorComponent } from '../../../editors/note-editor/terra-note-editor.component';
 import { TerraCodeEditorComponent } from '../../../editors/code-editor/terra-code-editor.component';
 import { TerraPortletComponent } from '../../../layouts/portlet/terra-portlet.component';
-import { TerraSuggestionBoxComponent } from '../../suggestion-box/terra-suggestion-box.component';
+import { SuggestionComponent } from '../form-components/suggestion/suggestion.component';
 import { DoubleInputComponent } from '../form-components/double-input/double-input.component';
 import { SliderComponent } from '../form-components/slider/slider.component';
 import { MultiSelectComponent } from '../form-components/multi-select/multi-select.component';
@@ -100,9 +100,9 @@ export class FormTypeMap {
 
     public readonly suggestion: TerraFormTypeInterface = {
         inputMap: {
-            required: 'inputIsRequired'
+            required: 'isRequired'
         },
-        component: TerraSuggestionBoxComponent
+        component: SuggestionComponent
     };
 
     // CONTAINER TYPES