+++ src/lib/terra-components.module.ts
@@ -17,6 +17,7 @@ import { DragDropModule } from '@angular/cdk/drag-drop';
 import { ModalModule } from 'ngx-bootstrap/modal';
 import { MatIconModule } from '@angular/material/icon';
 import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatSliderModule } from '@angular/material/slider';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
@@ -42,12 +43,13 @@ import { MatInputModule } from '@angular/material/input';
         DragDropModule,
         MatIconModule,
         MatCheckboxModule,
+        MatFormFieldModule,
+        MatInputModule,
+        MatAutocompleteModule,
         MatSliderModule,
         MatDatepickerModule,
         A11yModule,
-        MatFormFieldModule,
-        MatSelectModule,
-        MatInputModule
+        MatSelectModule
     ],
     declarations: [...components, ...directives, ...pipes],
     exports: [...exportedComponents, ...exportedDirectives, ...exportedPipes]