+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 -   **terra-form**
+    -   form type **suggestion** replaced deprecated `terra-suggestion-box` by a `mat-autocomplete` variant.
     -   form type **double** replaced deprecated `terra-double-input` by a `mat-input` variant.
     -   form type **slider** replaced deprecated `terra-slider` by a `mat-slider` variant.
     -   form type **multi select** replaced deprecated `tc-checkbox-group` by a `mat-select` variant in the multiple mode.